# src/api.yaml
openapi: 3.0.0
info:
  title: 'Melinda REST API for ILS integration '
  version: '2.3'
...
paths:
...
  /fix/{id}:
    post:
      summary: Run a fix operation on a single record
      tags:
        - /fix/
      parameters:
        - name: id
          description: The identifier of the record that's going to be updated by the fix operation
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/databaseIdNonEmpty'
        - name: fixType
          description: >-
            Fix operation to run. Usually available fixTypes:
            UNDEL for restoring a removed record,
            DELET for removing a record
          required: true
          in: query
          schema:
            type: string
        - name: cataloger
          description: >-
              Cataloger name to use while operating on a record. Note: requires admin rights to use
          in: query
          schema:
            type: string
        - name: noop
          description: Do not actually do the update the record.
          in: query
          schema:
            type: boolean
            default: false
      security:
        - httpBasic: []
      responses:
        '200':
          description: >-
            The record was succesfully updated. (Unless parameter **noop** was
            set to true, or parameter **skipNoChangeUpdate** was set to true and
            the update would have resulted in no changes in the datastore
            record.)
          headers:
            Record-ID:
              description: Identifier for the updated record in the datastore
              schema:
                $ref: '#/components/schemas/databaseIdNonEmpty'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/recordResponseItemFor200Update'
        '400':
          description: >-
            The request id invalid or the record data is malformed or has
            missmach content-type, or the parameters are invalid
        '401':
          description: Credentials are invalid or missing
        '403':
          description: >-
            The credentials are not authorized for this operation, i.e. the user
            does not have the permission to change the owner metadata of the
            record
        '404':
          description: The record does not exist
        '408':
          description: The request has timed out
        '409':
          description: The request contains out of date record data
        '415':
          description: The request is in a format that is not supported
        '422':
          description: The record does not pass validation
...
  /prio/:
    get:
      summary: Retrieve prio queue items
      tags:
        - /prio/
      parameters:
        - name: id
          description: Queue item identifier (correlationId)
          in: query
          required: false
          schema:
            type: string
      security:
        - httpBasic: []
      responses:
        '200':
          description: >-
            Array of queue item objects (See
            #/components/schemas/dbQueueItem-prio). Empty list if no resources
            were found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/dbQueueItem-prio'
        '401':
          description: Credentials are invalid or missing
        '403':
          description: The credentials are not authorized for this operation
        '404':
          description: The record does not exist
...
components:
  securitySchemes:
    httpBasic:
      type: http
      scheme: basic
  schemas:
...
    recordResponseItemFor200Update:
      type: object
      properties:
        databaseId:
          $ref: '#/components/schemas/databaseIdNonEmpty'
        recordMetadata:
          $ref: '#/components/schemas/recordMetadata'
        recordStatus:
          allOf:
            - example: UPDATED
            - $ref: '#/components/schemas/recordStatus'
        detailedRecordStatus:
          allOf:
            - example: UPDATED
            - $ref: '#/components/schemas/detailedRecordStatus'
        message:
          type: string
          example: Updated record 000123456
...
    databaseIdNonEmpty:
      allOf:
        - example: '000123456'
        - $ref: '#/components/schemas/databaseId'
    databaseId:
      type: string
      description: ID for a record in the datastore in nine-digit format
      pattern: ^\d{9}$
    recordStatus:
      description: Status for the handled record after the operation.
      type: string
      enum:
        - UPDATED
        - CREATED
        - SKIPPED
        - DUPLICATE
        - ERROR
        - INVALID
    detailedRecordStatus:
      description: More detailed status for handled record after the operation.
      type: string
      enum:
        - UPDATED
        - CREATED
        - SKIPPED
        - DUPLICATE
        - ERROR
        - INVALID
        - CONLICT
        - UNKNOWN
        - NOT_FOUND
...
    recordMetadata:
      description: Metadata extracted from the input record
      type: object
      properties:
        sourceIds:
          description: Record identifiers
          type: array
          items:
            type: string
            example: (FI-FOOBAR)123456
        title:
          description: Title for described resource
          type: string
          example: 'Hobitti :'
        standardIdentifiers:
          description: 'Standard identifiers for described resource: ISBNs, ISSNs etc.'
          type: array
          items:
            type: string
            example: 978-951-0-46873-9
        blobSequence:
          description: >-
            Sequence for the record in the input blob. Always "1" when handling
            single record input
          type: number
          example: 1
          pattern: ^\d+$
...
    queueItemState:
      type: string
      description: The request queue item state
      enum:
        - DONE
        - ERROR
        - IMPORTER.IN_PROCESS
        - IMPORTER.IN_QUEUE
        - IMPORTER.IMPORTING
        - VALIDATOR.PENDING_QUEUING
        - VALIDATOR.QUEUING_IN_PROGRESS
        - VALIDATOR.UPLOADING
        - VALIDATOR.PENDING_VALIDATION
        - VALIDATOR.VALIDATING
...
    dbQueueItem-prio:
      type: object
      properties:
        correlationId:
          type: string
          description: The identifier of the queue item
        cataloger:
          type: string
          description: The identifier of cataloger
        oCatalogerIn:
          type: string
          description: The identifier of cataloger used in actually saving the record
        operation:
          type: string
          description: The request operation type
          enum:
            - CREATE
            - UPDATE
        operationSettings:
          type: object
          properties:
            prio:
              type: boolean
              description: The operation is a prio operation
            unique:
              type: boolean
              description: >-
                Is the create operation a unique operation. If true, it does not
                create the record if there are duplicates in the datastore.
            noop:
              type: boolean
              description: >-
                Is the operation a noop operation. If true it will not not
                actually create or update the record in the datastore.
        queueItemState:
          $ref: '#/components/schemas/queueItemState'
        creationTime:
          type: string
          description: The request creation time
        modificationTime:
          type: string
          description: Last modification time of the request
        handledIds:
          type: array
          items:
            type: string
            description: Handled record id
        rejectedIds:
          type: array
          items:
            type: string
            description: Rejected record id
        errorStatus:
          type: string
          description: httpError status for an errrored job
        errorMessages:
          type: string
          description: Error message for an errrored job
        loadProcessReports:
          description: Array of load process reports for job.
          type: array
          items:
            description: >-
              A load process report. As default only kept for those processes
              that didn't handle all the records as expected
            type: object
            properties:
              status:
                type: string
                description: Http status from load process
              processId:
                type: string
                description: Process id of the load process
              processedAll:
                type: boolean
                description: Did load process process all the records
              recordAmount:
                type: number
                description: Number of records given to the load process
              processedAmount:
                type: number
                description: >-
                  Number of records processed (handled or rejected) by the load
                  process
              handledAmount:
                type: number
                description: >-
                  Number of records handled (updated or created) by the load
                  process
              rejectedAmount:
                type: number
                description: Number of records rejected by the load process
              rejectMessages:
                type: array
                description: >-
                  Reject messages given by the load process for the rejected
                  records
                items:
                  type: string
                  description: >-
                    A reject message given by the load process for a single
                    rejected record
        noopValidationMessages:
          type: array
          description: >-
            An array of validation messages given by validation to the record in
            a noop operation
          items:
            type: object
...